feat: extract animated phrase banner into standalone component below hero
